+++ native_library/android/FirebaseANE/VisionFace/src/main/java/com/tuarua/firebase/VisionFaceANE.java
@@ -26,6 +26,7 @@ public class VisionFaceANE implements FREExtension {
             ,"createGUID"
             ,"detect"
             ,"getResults"
+            ,"close"
     };
     private static VisionFaceANEContext extensionContext;