+++ native_library/apple/FirebaseANE/VisionBarcodeANE/VisionBarcodeANE.m
@@ -35,9 +35,8 @@ CONTEXT_INIT(TRFBVBC) {
         ,MAP_FUNCTION(TRFBVBC, createGUID)
         ,MAP_FUNCTION(TRFBVBC, detect)
         ,MAP_FUNCTION(TRFBVBC, getResults)
-        ,MAP_FUNCTION(TRFBVBC, isCameraSupported)
         ,MAP_FUNCTION(TRFBVBC, inputFromCamera)
-        ,MAP_FUNCTION(TRFBVBC, closeCamera)
+        ,MAP_FUNCTION(TRFBVBC, close)
     };
     SET_FUNCTIONS